+++ src/main/java/de/jeff_media/AngelChest/AngelChestHolder.java
@@ -1,0 +1,19 @@
+package de.jeff_media.AngelChest;
+
+import org.bukkit.inventory.Inventory;
+import org.bukkit.inventory.InventoryHolder;
+import org.jetbrains.annotations.NotNull;
+
+public class AngelChestHolder implements InventoryHolder {
+
+    Inventory inv;
+
+    void setInventory(Inventory inv) {
+        this.inv=inv;
+    }
+
+    @Override
+    public @NotNull Inventory getInventory() {
+        return inv;
+    }
+}