+++ src/main/java/de/jeff_media/AngelChest/CommandFetch.java
@@ -1,5 +1,7 @@
 package de.jeff_media.AngelChest;
 
+import org.bukkit.Bukkit;
+import org.bukkit.ChatColor;
 import org.bukkit.Location;
 import org.bukkit.block.Block;
 import org.bukkit.block.BlockFace;
@@ -18,33 +20,48 @@ public class CommandFetch implements CommandExecutor {
 
 	@Override
 	public boolean onCommand(CommandSender sender, Command command, String alias, String[] args) {
-		if(!command.getName().equalsIgnoreCase("acfetch")) return false;		
 
-		if(!(sender instanceof Player)) {
+		Player affectedPlayer=null;
+
+		if(!command.getName().equalsIgnoreCase("acfetch")) return false;
+
+		if(!(sender instanceof Player) && args.length==0) {
 			sender.sendMessage(plugin.messages.MSG_PLAYERSONLY);
 			return true;
 		}
 
-		Player p = (Player) sender;
+		if(args.length>1 && sender.hasPermission("angelchest.others")) {
+
+			Player p = Bukkit.getPlayer(args[1]);
+			if(p==null) {
+				sender.sendMessage(ChatColor.RED+"Could not find player "+args[1]);
+				return true;
+			}
+
+			affectedPlayer = Bukkit.getPlayer(args[1]);
+		}
+
+		Player sendTo = (Player) sender;
+		if(affectedPlayer==null) affectedPlayer=sendTo;
 
 		if(!sender.hasPermission("angelchest.fetch")) {
 			sender.sendMessage(plugin.getCommand("acfetch").getPermissionMessage());
 			return true;
 		}
 
-        AngelChest ac = AngelChestCommandUtils.argIdx2AngelChest(plugin, p, args);
+        AngelChest ac = AngelChestCommandUtils.argIdx2AngelChest(plugin, sendTo, affectedPlayer, args);
         if(ac == null) {
             return true;
         }
 
         // Only charge the player if they select a valid chest
 		double price = plugin.getConfig().getDouble("price-fetch");
-		if(price>0 && !AngelChestCommandUtils.hasEnoughMoney(p,price,plugin)) {
+		if(price>0 && !AngelChestCommandUtils.hasEnoughMoney(sendTo,price,plugin)) {
 			return true;
 		}
 
-		Location newLoc = p.getLocation();
-		String dir = Utils.getCardinalDirection(p);
+		Location newLoc = sendTo.getLocation();
+		String dir = Utils.getCardinalDirection(sendTo);
 		BlockFace facing;
 	
 		// Set the relative direction of the block and offset the new chest location
@@ -91,7 +108,7 @@ public class CommandFetch implements CommandExecutor {
 
 		// Move the block in game
 		ac.destroyChest(oldBlock);
-		ac.createChest(newBlock, p.getUniqueId());
+		ac.createChest(newBlock, sendTo.getUniqueId());
 
 		// Make the chest face the player
 		try {
@@ -104,7 +121,7 @@ public class CommandFetch implements CommandExecutor {
 		plugin.angelChests.put(newBlock, plugin.angelChests.remove(oldBlock));
 		plugin.angelChests.get(newBlock).block = newBlock;
 
-		p.sendMessage("AngelChest Retrieved!");
+		sendTo.sendMessage("AngelChest Retrieved!"); // TODO: Configurable
 
 		return true;
 	}