+++ src/main/java/de/jeff_media/AngelChest/CommandTeleportTo.java
@@ -1,5 +1,7 @@
 package de.jeff_media.AngelChest;
 
+import org.bukkit.Bukkit;
+import org.bukkit.ChatColor;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
@@ -15,21 +17,35 @@ public class CommandTeleportTo implements CommandExecutor {
 
 	@Override
 	public boolean onCommand(CommandSender sender, Command command, String alias, String[] args) {
-		if(!command.getName().equalsIgnoreCase("actp")) return false;		
+		Player affectedPlayer = null;
+		if(!command.getName().equalsIgnoreCase("actp")) return false;
 
-		if(!(sender instanceof Player)) {
+		if(!(sender instanceof Player) && args.length==0) {
 			sender.sendMessage(plugin.messages.MSG_PLAYERSONLY);
 			return true;
 		}
 
+		if(args.length>1 && sender.hasPermission("angelchest.others")) {
+
+			Player p = Bukkit.getPlayer(args[1]);
+			if(p==null) {
+				sender.sendMessage(ChatColor.RED+"Could not find player "+args[1]);
+				return true;
+			}
+
+			affectedPlayer = Bukkit.getPlayer(args[1]);
+
+		}
+
 		Player p = (Player) sender;
+		if(affectedPlayer==null) affectedPlayer=p;
 
 		if(!sender.hasPermission("angelchest.tp")) {
 			sender.sendMessage(plugin.getCommand("actp").getPermissionMessage());
 			return true;
 		}
 
-        AngelChest ac = AngelChestCommandUtils.argIdx2AngelChest(plugin, p, args);
+        AngelChest ac = AngelChestCommandUtils.argIdx2AngelChest(plugin, p, affectedPlayer, args);
         if(ac == null) {
             return true;
         }