+++ src/main/java/de/jeff_media/AngelChest/ConfigUtils.java
@@ -31,6 +31,7 @@ public class ConfigUtils {
 		File groupsFile = new File(main.getDataFolder()+File.separator+"groups.yml");
 		main.groupUtils = new GroupUtils(main,groupsFile);
 		main.worldGuardHandler = new WorldGuardHandler(main);
+		main.hookUtils = new HookUtils(main);
 		if(reload) {
 			main.loadAllAngelChestsFromFile();
 		}
@@ -68,6 +69,10 @@ public class ConfigUtils {
 		main.getConfig().addDefault("disable-worldguard-integration",false);
 		main.getConfig().addDefault("ignore-keep-inventory",false);
 		main.getConfig().addDefault("head-uses-player-name",true);
+		main.getConfig().addDefault("auto-respawn",false);
+		main.getConfig().addDefault("auto-respawn-delay",10);
+		main.getConfig().addDefault("play-can-skip-auto-respawn",false);
+		main.getConfig().addDefault("use-slimefun",true);
 		main.getConfig().addDefault("full-xp", false); // Currently not in config because there is no way to get players XP
 		main.disabledWorlds = (ArrayList<String>) main.getConfig().getStringList("disabled-worlds");
 		main.disabledRegions = (ArrayList<String>) main.getConfig().getStringList("disabled-worldguard-regions");