+++ src/main/java/de/jeff_media/AngelChest/HookUtils.java
@@ -2,33 +2,31 @@ package de.jeff_media.AngelChest;
 
 import java.util.Objects;
 
-import org.bukkit.ChatColor;
+import io.github.thebusybiscuit.slimefun4.utils.SlimefunUtils;
 import org.bukkit.inventory.ItemStack;
-
-import io.github.thebusybiscuit.slimefun4.implementation.items.magical.SoulboundItem;
-import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
 
 public class HookUtils {
 
-    static final String SOULBOUND_LORE= ChatColor.GRAY+"Soulbound";
+    Main main;
 
-    static boolean isSlimefunSoulbound(ItemStack item) {
-
-
-
-        if(item==null) return false;
-        if(!item.hasItemMeta()) return false;
-        if(!item.getItemMeta().hasLore()) return false;
-        for(String lore : item.getItemMeta().getLore()) {
-            if(lore.equals(SOULBOUND_LORE)) {
-                return true;
-            }
-        }
-        return false;
-
+    HookUtils(Main main) {
+        this.main=main;
     }
 
-    static boolean keepOnDeath(ItemStack item) {
+    boolean isSlimefunSoulbound(ItemStack item) {
+        if(item==null) return false;
+        if(!main.getConfig().getBoolean("use-slimefun")) return false;
+
+        try {
+            Class.forName("io.github.thebusybiscuit.slimefun4.utils.SlimefunUtils");
+            return SlimefunUtils.isSoulbound(item);
+        } catch (ClassNotFoundException | NoClassDefFoundError e) {
+            main.getConfig().set("use-slimefun",false);
+            return false;
+        }
+    }
+
+    boolean keepOnDeath(ItemStack item) {
         if(item==null) return false;
         return isSlimefunSoulbound(item);
     }